Refactor basic components to seperate directory
