+++ internals/LED_data/Components/_RGB.py
