+++ internals/LED_data/Components/_RGBArray.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from rpi_ws281x import PixelStrip
 
 # from ._Animation.Animation import Animation
-from .RGB import RGB
+from ._RGB import RGB
 
 class RGBArray():
     def __init__(self: RGBArray, array_length: int):
