+++ internals/LED_data/__init__.py
@@ -1,13 +1,12 @@
+from .Components import Animation, RGB, RGBArray
+
 from ._Animation.AlternatingAnimation import AlternatingAnimation
-from ._Animation.Animation import Animation
 from . import color_constants
 from ._Animation.FillAnimation import FillAnimation
 from ._Interrupt.FlashAnimation import FlashAnimation
 from ._Interrupt.InterruptAnimation import InterruptAnimation
 from ._Animation.KillAnimation import KillAnimation
 from ._Animation.PulseAnimation import PulseAnimation
-from .RGB import RGB
-from .RGBArray import RGBArray
 from ._Animation.TrafficAnimation import TrafficAnimation
 from ._Animation.TransitionAnimation import TransitionAnimation
 from ._Animation.WaveAnimation import WaveAnimation
